+++ tests/testthat/test-visualization_functions.R
@@ -2,7 +2,7 @@
 ## Project: pathfindR
 ## Script purpose: Testthat testing script for
 ## visualization-related functions
-## Date: Jan 16, 2022
+## Date: Feb 7, 2022
 ## Author: Ege Ulgen
 ##################################################
 
@@ -420,6 +420,9 @@ test_that("`term_gene_heatmap()` produces a ggplot object using the correct data
 
   # genes_df supplied - wihout change column
   expect_is(p <- term_gene_heatmap(RA_output[1:3, ], RA_input[, -2]), "ggplot")
+
+  # sort by lowest_p instead
+  expect_is(p <- term_gene_heatmap(RA_output[1:3, ], RA_input, sort_terms_by_p = TRUE), "ggplot")
 })
 
 test_that("`term_gene_graph()` arg checks work", {